Server/requirements.txt
# requirements.txt
# For the Server
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
